+++ tests/maxo_dialog/widgets/input/test_combined.py
@@ -1,23 +1,10 @@
-from datetime import UTC, datetime
 from unittest.mock import AsyncMock
 
 from maxo.dialogs import DialogManager
 from maxo.dialogs.widgets.input.combined import CombinedInput
 from maxo.dialogs.widgets.input.text import TextInput
-from maxo.enums import ChatType
-from maxo.routing.updates import MessageCreated
-from maxo.types import Message, MessageBody, Recipient
 
-
-def create_message(text: str) -> MessageCreated:
-    return MessageCreated(
-        timestamp=datetime(2024, 1, 1, tzinfo=UTC),
-        message=Message(
-            timestamp=datetime(2024, 1, 1, tzinfo=UTC),
-            recipient=Recipient(chat_type=ChatType.DIALOG, user_id=1),
-            body=MessageBody(mid="test_mid", seq=1, text=text),
-        ),
-    )
+from .conftest import create_text_message
 
 
 async def test_combined_input_first_accepts(mock_manager: DialogManager) -> None:
@@ -27,7 +14,7 @@ async def test_combined_input_first_accepts(mock_manager: DialogManager) -> None
     input2 = TextInput(id="text2", on_success=on_success2)
     combined = CombinedInput(input1, input2)
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is True
@@ -47,7 +34,7 @@ async def test_combined_input_second_accepts(mock_manager: DialogManager) -> Non
     input2 = TextInput(id="text2", on_success=on_success2)
     combined = CombinedInput(input1, input2)
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is True
@@ -64,7 +51,7 @@ async def test_combined_input_all_reject(mock_manager: DialogManager) -> None:
     input2 = TextInput(id="text2", on_success=on_success2, filter=filter2)
     combined = CombinedInput(input1, input2)
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is False
@@ -78,7 +65,7 @@ async def test_combined_input_with_filter_accepts(mock_manager: DialogManager) -
     filter_func = AsyncMock(return_value=True)
     combined = CombinedInput(input1, filter=filter_func)
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is True
@@ -92,7 +79,7 @@ async def test_combined_input_with_filter_rejects(mock_manager: DialogManager) -
     filter_func = AsyncMock(return_value=False)
     combined = CombinedInput(input1, filter=filter_func)
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is False
@@ -103,7 +90,7 @@ async def test_combined_input_with_filter_rejects(mock_manager: DialogManager) -
 async def test_combined_input_no_inputs(mock_manager: DialogManager) -> None:
     combined = CombinedInput()
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is False
@@ -126,7 +113,7 @@ async def test_combined_input_three_inputs(mock_manager: DialogManager) -> None:
     input3 = TextInput(id="text3", on_success=on_success3)
     combined = CombinedInput(input1, input2, input3)
 
-    message = create_message("Test")
+    message = create_text_message("Test")
     result = await combined.process_message(message, None, mock_manager)
 
     assert result is True
@@ -142,7 +129,7 @@ async def test_combined_input_empty_message(mock_manager: DialogManager) -> None
     input2 = TextInput(id="text2", on_success=on_success2)
     combined = CombinedInput(input1, input2)
 
-    message = create_message("")
+    message = create_text_message("")
     result = await combined.process_message(message, None, mock_manager)
 
     # Both TextInput widgets reject empty messages
